lmstudio-agent/requirements.txt
lmstudio==1.5.0
requests>=2.32.2
pandas==2.1.4
openpyxl>=3.1.5
python-docx==1.1.0
